nfapi: fix integer overflows in SRS report handling for large arrays
With 64 gNB antenna elements the SRS channel IQ matrix exceeds the
ranges the SRS report code path was written for (272 PRGs * 2 UE ports
* 64 elements * 4 B = 136 KiB per report):

- unpack_nr_srs_report_tlv_value(): last_idx was int16_t, which
  overflows at report lengths >= 128 KiB, so the copy loop never ran
  and the report was silently dropped. Widen to int32_t and reject
  reports larger than the value buffer instead of overrunning it.
- pack/unpack_nr_srs_normalized_channel_iq_matrix():
  channel_matrix_size was uint16_t and wraps at 64 KiB, truncating the
  matrix. Widen to uint32_t and bound it by sizeof(channel_matrix).
- handle_nr_srs_measurements(): the SRS_IND_DEBUG print indexed the
  matrix with a uint16_t, which wraps for Nu*Ng*Np > 65535.

+++ openair2/LAYER2/NR_MAC_gNB/gNB_scheduler_ulsch.c
@@ -1563,7 +1563,7 @@ void handle_nr_srs_measurements(const module_id_t module_id,
       for (int uI = 0; uI < nr_srs_channel_iq_matrix.num_ue_srs_ports; uI++) {
         for (int gI = 0; gI < nr_srs_channel_iq_matrix.num_gnb_antenna_elements; gI++) {
           for (int pI = 0; pI < nr_srs_channel_iq_matrix.num_prgs; pI++) {
-            uint16_t index = uI * nr_srs_channel_iq_matrix.num_gnb_antenna_elements * nr_srs_channel_iq_matrix.num_prgs + gI * nr_srs_channel_iq_matrix.num_prgs + pI;
+            uint32_t index = (uint32_t)uI * nr_srs_channel_iq_matrix.num_gnb_antenna_elements * nr_srs_channel_iq_matrix.num_prgs + gI * nr_srs_channel_iq_matrix.num_prgs + pI;
             LOG_I(NR_MAC,
                   "(uI %i, gI %i, pI %i) channel_matrix --> real %i, imag %i\n",
                   uI,
